Oops checkin fix for accidentally dropping the 'size_t' version here
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10175 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/NclGRIB.c
@@ -101,7 +101,7 @@ static PtableInfo *Ptables = NULL;
 
 extern void GribPushAtt(
 #if NhlNeedProto
-GribAttInqRecList **att_list_ptr,char* name,void *val,size_t dimsize,NclObjClass type
+GribAttInqRecList **att_list_ptr,char* name,void *val,int dimsize,NclObjClass type
 #endif
 );
 
@@ -1307,7 +1307,7 @@ GribFileRecord *therec;
 	GribParamList *step = NULL;
 	NclQuark *tmp_string = NULL;
 	int *tmp_int = NULL;
-	size_t tmp_dimsizes = 1;
+	int tmp_dimsizes = 1;
 	GribRecordInqRec *grib_rec = NULL;
 	GribAttInqRecList *att_list_ptr= NULL;
 	int i;
@@ -1841,7 +1841,7 @@ float bytes2float (unsigned char *bytes)
 					
 
 void _Do109(GribFileRecord *therec,GribParamList *step) {
-	size_t dimsizes_level;
+	int dimsizes_level;
 	int tmp_file_dim_number;
 	int i;
 	char buffer[256];
@@ -1857,7 +1857,7 @@ void _Do109(GribFileRecord *therec,GribParamList *step) {
 	int sign;
 	float tmpb;
 	float tmpa;
-	size_t count;
+	int count;
 	int interface = 0;
 	GribAttInqRecList *att_list_ptr= NULL;	
 	int attcount;
@@ -2375,7 +2375,7 @@ GribFileRecord *therec;
 
 	step = therec->it_dims;
 	for (i = 0; i < therec->n_it_dims; i++) {
-		size_t dimsize;
+		int dimsize;
 		NrmQuark *vals;
 		double *dates;
 		double *hours;
@@ -2465,9 +2465,9 @@ GribFileRecord *therec;
 	int n_dims_lon = 0;
 	int n_dims_rot = 0;
 	int n_dims_level = 0;
-	size_t *dimsizes_lat = NULL;
-	size_t *dimsizes_lon = NULL;
-	size_t *dimsizes_rot = NULL;
+	int *dimsizes_lat = NULL;
+	int *dimsizes_lon = NULL;
+	int *dimsizes_rot = NULL;
 	float *tmp_lat = NULL;
 	float *tmp_lon = NULL;
 	float *tmp_rot = NULL;
@@ -2480,8 +2480,8 @@ GribFileRecord *therec;
 	int natts = 0;
 	NclQuark *tmp_string = NULL;
 	float *tmp_float = NULL;
-	size_t tmp_dimsizes = 1;
-	size_t dimsizes_level = 1;
+	int tmp_dimsizes = 1;
+	int dimsizes_level = 1;
 	int nlonatts = 0;
 	int nlatatts = 0;
 	int nrotatts = 0;
@@ -3858,7 +3858,7 @@ GribParamList* step;
 	int n_tmp_ens_vals = 0;
 	ENS *tmp_ens_vals;
 /*	int total;*/
-	size_t total, tmp_dim_siz = 0;
+	int total, tmp_dim_siz = 0;
 	int doff;
 	char *name;
 	float *lprob, *uprob;
@@ -4028,7 +4028,7 @@ GribParamList* step;
 		step->upper_probs = NULL;
 		step->probability = NULL;
 		if (tmp_ens_vals[0].prob_type == 1) {
-            tmp_dim_siz = (size_t) n_tmp_ens_vals;
+            tmp_dim_siz = (int) n_tmp_ens_vals;
 			step->probability = (NclOneDValCoordData)_NclCreateVal(
 				NULL,
 				NULL,
@@ -4045,7 +4045,7 @@ GribParamList* step;
 			NclFree(uprob);
 		}
 		else if (tmp_ens_vals[0].prob_type == 2) {
-            tmp_dim_siz = (size_t) n_tmp_ens_vals;
+            tmp_dim_siz = (int) n_tmp_ens_vals;
 			step->probability = (NclOneDValCoordData)_NclCreateVal(
 				NULL,
 				NULL,
@@ -4062,7 +4062,7 @@ GribParamList* step;
 			NclFree(lprob);
 		}
 		else {
-            tmp_dim_siz = (size_t) n_tmp_ens_vals;
+            tmp_dim_siz = (int) n_tmp_ens_vals;
 			step->lower_probs = (NclMultiDValData)_NclCreateVal(
 				NULL,
 				NULL,
@@ -4098,7 +4098,7 @@ GribParamList* step;
 			ens_vals_q[j] = GetEnsQuark(&(tmp_ens_vals[j]));
 			ens_indexes[j] = j;
 		}
-        tmp_dim_siz = (size_t) n_tmp_ens_vals;
+        tmp_dim_siz = (int) n_tmp_ens_vals;
 		step->ensemble = (NclMultiDValData)_NclCreateVal(
 					NULL,
 					NULL,
@@ -4145,8 +4145,8 @@ GribParamList* step;
 			it_vals_q[j] = GetItQuark(&(tmp_it_vals[j]));
 	}
 	if(n_tmp_it_vals > 1 || (n_tmp_it_vals > 0 && (therec->single_dims & GRIB_Initial_Time_Dims))) {
-		step->var_info.dim_sizes[i] = (size_t) n_tmp_it_vals;
-        tmp_dim_siz = (size_t) n_tmp_it_vals;
+		step->var_info.dim_sizes[i] = (int) n_tmp_it_vals;
+        tmp_dim_siz = (int) n_tmp_it_vals;
 		step->yymmddhh = (NclOneDValCoordData)_NclCreateVal(
 					NULL,
 					NULL,
@@ -4166,7 +4166,7 @@ GribParamList* step;
 		i++;
 	} else if(n_tmp_it_vals == 1) {
 		step->yymmddhh_isatt = 1;
-        tmp_dim_siz = (size_t) n_tmp_it_vals;
+        tmp_dim_siz = (int) n_tmp_it_vals;
 		step->yymmddhh = (NclOneDValCoordData)_NclCreateVal(
 					NULL,
 					NULL,
@@ -4190,8 +4190,8 @@ GribParamList* step;
 */
 	}
 	if(n_tmp_ft_vals > 1 || (n_tmp_ft_vals > 0 && (therec->single_dims & GRIB_Forecast_Time_Dims))) {
-		step->var_info.dim_sizes[i] = (size_t) n_tmp_ft_vals;
-        tmp_dim_siz = (size_t) n_tmp_ft_vals;
+		step->var_info.dim_sizes[i] = (int) n_tmp_ft_vals;
+        tmp_dim_siz = (int) n_tmp_ft_vals;
 		step->forecast_time = (NclOneDValCoordData)_NclCreateVal(
 					NULL,
 					NULL,
@@ -4208,7 +4208,7 @@ GribParamList* step;
 		step->forecast_time_isatt = 0;
 		i++;
 	} else if(n_tmp_ft_vals == 1) {
-        tmp_dim_siz = (size_t) n_tmp_ft_vals;
+        tmp_dim_siz = (int) n_tmp_ft_vals;
 		step->forecast_time = (NclOneDValCoordData)_NclCreateVal(
 					NULL,
 					NULL,
@@ -4231,8 +4231,8 @@ GribParamList* step;
 	}
 	if((tmp_lv_vals != NULL)&&(tmp_lv_vals1 == NULL)) {
 		if(n_tmp_lv_vals > 1 || (n_tmp_lv_vals > 0 && (therec->single_dims & GRIB_Level_Dims))) {
-			step->var_info.dim_sizes[i] = (size_t) n_tmp_lv_vals;
-            tmp_dim_siz = (size_t) n_tmp_lv_vals;
+			step->var_info.dim_sizes[i] = (int) n_tmp_lv_vals;
+            tmp_dim_siz = (int) n_tmp_lv_vals;
 			step->levels = (NclOneDValCoordData)_NclCreateVal(
 						NULL,
 						NULL,
@@ -4252,7 +4252,7 @@ GribParamList* step;
 				step->levels_isatt = 0;
 		} else if (n_tmp_lv_vals == 1) {
 			step->levels_isatt = 1;
-            tmp_dim_siz = (size_t) n_tmp_lv_vals;
+            tmp_dim_siz = (int) n_tmp_lv_vals;
 			step->levels = (NclOneDValCoordData)_NclCreateVal(
 						NULL,
 						NULL,
@@ -4276,8 +4276,8 @@ GribParamList* step;
 		}
 	} else if((tmp_lv_vals != NULL)&&(tmp_lv_vals1 != NULL)) { 
 		if(n_tmp_lv_vals > 1 || (n_tmp_lv_vals > 0 && (therec->single_dims & GRIB_Level_Dims))) {
-			step->var_info.dim_sizes[i] = (size_t) n_tmp_lv_vals;
-            tmp_dim_siz = (size_t) n_tmp_lv_vals;
+			step->var_info.dim_sizes[i] = (int) n_tmp_lv_vals;
+            tmp_dim_siz = (int) n_tmp_lv_vals;
 			step->levels = NULL;
 			step->levels0 = (NclMultiDValData)_NclCreateVal(
 						NULL,
@@ -4311,7 +4311,7 @@ GribParamList* step;
 		} else if (n_tmp_lv_vals == 1) {
 			step->levels_isatt = 1;
 			step->levels = NULL;
-            tmp_dim_siz = (size_t) n_tmp_lv_vals;
+            tmp_dim_siz = (int) n_tmp_lv_vals;
 			step->levels0 = (NclMultiDValData)_NclCreateVal(
 						NULL,
 						NULL,
@@ -7644,7 +7644,7 @@ void* storage;
 	long *grid_start;
 	long *grid_finish;
 	long *grid_stride;
-/*	size_t n_other_dims;*/
+/*	int n_other_dims;*/
 	int n_other_dims;
 	int current_index[5] = {0,0,0,0,0};
 	int dim_offsets[5] = {-1,-1,-1,-1,-1};
@@ -7657,7 +7657,7 @@ void* storage;
 	NclScalar missingv;
 	int int_or_float;
 	int fd;
-	size_t grid_dim_sizes[3];
+	int grid_dim_sizes[3];
 	int n_grid_dims;
 	NclMultiDValData tmp_md;
 	NclSelectionRecord  sel_ptr;
